Backport of client/fingerprint: correctly fingerprint E/P cores of Apple Silicon chips into release/1.5.x (#16690)
This pull request was automerged via backport-assistant

+++ drivers/docker/util/stats_windows.go
@@ -42,7 +42,7 @@ func DockerStatsToTaskResourceUsage(s *docker.Stats) *cstructs.TaskResourceUsage
 		ThrottledPeriods: s.CPUStats.ThrottlingData.ThrottledPeriods,
 		ThrottledTime:    s.CPUStats.ThrottlingData.ThrottledTime,
 		Percent:          cpuPercent,
-		TotalTicks:       (cpuPercent / 100) * stats.TotalTicksAvailable() / float64(runtime.NumCPU()),
+		TotalTicks:       (cpuPercent / 100) * float64(stats.TotalTicksAvailable()) / float64(runtime.NumCPU()),
 		Measured:         DockerMeasuredCPUStats,
 	}
